[TST] Add workaround for partial tests

+++ tests/test_utils.py
@@ -18,8 +18,7 @@ class TestUtils(unittest.TestCase):
     
     @classmethod
     def tearDownClass(cls):
-        # missing_ok=True is only available in python 3.8+
-        remove_file = lambda file: Path(file).unlink()
+        remove_file = lambda file: Path(file).unlink() if Path(file).exists() else None
         remove_dir = lambda dir: shutil.rmtree(dir, ignore_errors=True)
 
         remove_file(cls.tiny_evil_file)
